Wellbeing RAG/requirements.txt
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
build==1.2.1
cachetools==5.3.3
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fsspec==2024.6.0
google-auth==2.30.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==30.1.0
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.2.13
langchain-openai==0.1.15
langchainhub==0.1.14
langsmith==0.1.85
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.35.13
opentelemetry-api==1.25.0
overrides==7.7.0
packaging==23.2
pinecone-client==4.1.2
pinecone-notebooks==0.1.1
pinecone-plugin-interface==0.0.7
pulsar-client==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pypdf==4.2.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
